Maximally reduce fmpq_poly.h
