+++ src/fmpq_poly/comparisons.c
@@ -1,0 +1,23 @@
+/*
+    Copyright (C) 2023 Albin Ahlbäck
+
+    This file is part of FLINT.
+
+    FLINT is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 2.1 of the License, or
+    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
+*/
+
+#include "fmpz.h"
+#include "fmpq_poly.h"
+
+int fmpq_poly_is_zero(const fmpq_poly_t poly)
+{
+    return poly->length == WORD(0);
+}
+
+int fmpq_poly_is_one(const fmpq_poly_t poly)
+{
+    return (poly->length == WORD(1)) && (fmpz_equal(poly->coeffs, poly->den));
+}
